+++ app/views/users/profile.html.erb
@@ -2,7 +2,7 @@
 
 
 <h3>アイコン画像</h3>
-<%# <%= @user.img.variant(resize: '500x500') %> 
+<%= @user.img.variant(resize: '500x500') %>
 
 
 <h3>名前</h3>
@@ -11,6 +11,6 @@
 <h3>自己紹介</h3>
 <%= @user.introduction %>
 
-<%= link_to "編集", users_profile_edit_path, class: 'post' %>
+<%= link_to "編集", profiles_edit_path, class: 'post' %>
 
 <%= link_to "アカウント", users_account_path, class: 'post' %>
