+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users, controllers: {
     registrations: 'users/registrations'
-}
+  }
   root 'top#index'
   devise_scope :user do
     get 'users/index'
@@ -9,20 +9,24 @@ Rails.application.routes.draw do
     get '/users/sign_up' => 'devise/registrations#new'
     get '/users', to: redirect("/users/sign_up")
   end
-    get "users/profile" => "users#profile"
-    get "users/profile/edit" => "users#edit"
-    get "users/profile/edit" => "users#update"
-    get "users/account" => "users#account"
-    get "users/account" => "users#account"
-    get "/users" => "users#edit"
-    get "/users" => "users#update"
-    # get "rooms/index"
-    get "rooms/new" => "rooms#new"
-    # get "rooms/own" => "rooms#own"
-    # get "/users" => "users#update"
-
-  resources :users
-  resource :profile,only: %i[show edit update]
+  get "users/profile" => "users#profile"
+  get "users/profile/edit" => "users#edit"
+  get "users/profile/edit" => "profiles#edit"
+  get "users/profile/edit" => "users#update"
+  get "users/profile/edit" => "profiles#update"
+  get "users/account" => "users#account"
+  get "users/account" => "users#account"
+  get "/users" => "users#edit"
+  get "/users" => "users#update"
+  # get "rooms/index"
+  get "rooms/new" => "rooms#new"
+  get "rooms/new" => "rooms#create"
+  # get "rooms/own" => "rooms#own"
+  # get "/users" => "users#update"
+  
+  # resources :users
+  resources :rooms
+  resource :profiles,only: %i[show edit update]
     # sessions: "users/sessions",
     # registrations: "users/registrations"
